(bugfix): fix compiled code and unresolve require calls in transient dependencies
In the 'stateful.runme' extension we experienced issues with trying to bundle
this package due to some transient dependencies trying to load files from
the node_modules directory. This patch removes these calls with a NOOP. This
can potentially influence the way the package work as it is unknown how these
dependencies were used.